# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Airplane Museum API",
    "description": "API for managing airplanes and museums",
    "version": "1.0.0"
  },
  "host": "airplane-museum-api.onrender.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/auth/status": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/auth/test-protected": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/auth/logout": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}